Parser completed

+++ headerCollections/Courses.h
@@ -5,7 +5,7 @@ class Courses {
     private:
         string courseID;
         string courseName;
-        unsigned units;
+        double units;
         string days;
         string time;
         string location;
@@ -15,18 +15,18 @@ class Courses {
         //Public Constructors
         Courses();
         Courses(string courseName);
-        Courses(string courseID, string courseName, unsigned units, string days, string time, string location);
+        Courses(string courseID, string courseName, double units, string days, string time, string location);
 
         void changeCourseID(string newCourseID);
         void changeCourseName(string newCourseName);
-        void changeUnitsCtr(int newUnitCtr);
+        void changeUnitsCtr(double newUnitCtr);
         void changeDays(string newDays);
         void changeTimes(string newTimes);
         void changeLocations(string newLocations);
 
         string getCourseID();
         string getCourseName();
-        unsigned getUnitsCtr();
+        double getUnitsCtr();
         string getDays();
         string getTimes();
         string getLocations();
